feat: Merge InputBox tests, reduce to smoke tests, and remove Vitest worker config
