+++ packages/code/vitest.config.ts
@@ -25,7 +25,6 @@ export default defineConfig(() => {
         FILE_SELECTOR_DEBOUNCE_MS: "0",
         PASTE_DEBOUNCE_MS: "0",
       },
-      maxWorkers: isCI ? 1 : 3,
     },
     resolve: {
       alias: {
